+++ src/project/contract/eravm_assembly.rs
@@ -5,6 +5,7 @@
 use crate::build::contract::Contract as ContractBuild;
 use crate::message_type::MessageType;
 use crate::project::contract::metadata::Metadata as ContractMetadata;
+use crate::vyper::selection::Selection as VyperSelection;
 
 ///
 /// The EraVM assembly contract.
@@ -37,7 +38,7 @@ impl Contract {
         source_code_hash: Option<[u8; era_compiler_common::BYTE_LENGTH_FIELD]>,
         optimizer_settings: era_compiler_llvm_context::OptimizerSettings,
         llvm_options: Vec<String>,
-        output_assembly: bool,
+        output_selection: Vec<VyperSelection>,
         _suppressed_messages: Vec<MessageType>,
         debug_config: Option<era_compiler_llvm_context::DebugConfig>,
     ) -> anyhow::Result<ContractBuild> {
@@ -57,18 +58,10 @@ impl Contract {
             contract_path,
             self.source_code,
             metadata_hash,
-            output_assembly,
+            output_selection.contains(&VyperSelection::EraVMAssembly),
             debug_config.as_ref(),
         )?;
 
-        Ok(ContractBuild::new(
-            build,
-            None,
-            None,
-            None,
-            None,
-            None,
-            vec![],
-        ))
+        Ok(ContractBuild::new_inner(build))
     }
 }